# docs/swagger/atlas-api.yaml
...
openapi: 3.0.3
info:
  title: ScaleMargin Dispatcher API — Atlas data plane
  version: "1.0.0"
  description: |
    The HTTP API the **ScaleMargin Atlas** platform calls to render a dispatcher
    dashboard, author its variables, and read campaign reporting and logs.

    **The contract in one line:** the base URL changes per client, the endpoints
    never do. Atlas hard-codes the paths and varies only the host.

    ## Two routers

    | Tree | Audience | Auth | Reachability |
    | --- | --- | --- | --- |
    | `/api/v1/data-plane/*` | Atlas | API key | must be reachable from ScaleMargin |
    | `/api/v1/internal/*` | the client's own infrastructure | none | should **not** be internet-exposed |

    ## Authentication

    One environment variable on the dispatcher (`DISPATCHER_ATLAS_KEY`, 32+
    chars, e.g. `openssl rand -base64 32`). No database record, no UI — the
    deployment owns the credential.

    * Send it as `Authorization: Bearer <key>`. **Never in a query string** —
      it lands in access logs, proxy logs and browser history.
    * Comparison is constant-time over SHA-256 digests; neither the key's length
      nor its prefix leaks through response timing.
    * **Rotation/revocation = change the value and restart.** No runtime kill switch.
    * **Unset means disabled:** every data-plane route returns `503 unavailable`
      while `DISPATCHER_ATLAS_KEY` is absent — including requests presenting a
      key. It fails closed and never falls open.
...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/data-plane/call-metadata:
    get:
      operationId: listCallMetadata
      summary: List call metadata schemas
      description: >
        Named sets of keys an api variable can attach. Each schema carries
        `used_by` — the api variables attaching it. Not paginated (tens per
        dispatcher).
      tags: [Call Metadata]
      responses:
        "200":
          description: Every schema
          content:
            application/json:
              schema:
                type: object
                required: [generated_at, call_metadata]
                properties:
                  generated_at: { type: string, format: date-time }
                  call_metadata:
                    type: array
                    items: { $ref: "#/components/schemas/CallMetadata" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
    post:
      operationId: createCallMetadata
      summary: Create a call metadata schema
      tags: [Call Metadata]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CallMetadataInput" }
            example:
              name: gold_metadata
              keys:
                - { key: tenure, placeholder: "24", regex: "^\\d+$" }
                - { key: age, placeholder: "36" }
                - { key: limit }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_metadata: { $ref: "#/components/schemas/CallMetadata" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/data-plane/call-metadata/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string, pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$" }
    get:
      operationId: getCallMetadata
      summary: Get one call metadata schema
      tags: [Call Metadata]
      responses:
        "200":
          description: The schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_metadata: { $ref: "#/components/schemas/CallMetadata" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateCallMetadata
      summary: Rename and/or replace the key list
      description: >
        Send `name`, `keys`, or both; `keys` replaces the list whole. Renaming is
        always safe — api variables reference the schema by id. Removing or
        renaming a key an api variable's request still uses is **409 conflict**,
        with one `details[]` entry (path `keys`) per variable and key.
      tags: [Call Metadata]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                keys:
                  type: array
                  items: { $ref: "#/components/schemas/CallMetadataKey" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_metadata: { $ref: "#/components/schemas/CallMetadata" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      operationId: deleteCallMetadata
      summary: Delete a call metadata schema
      description: Refused with **409 conflict** while any api variable attaches it; the message names them.
      tags: [Call Metadata]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
                  name: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        The dispatcher's DISPATCHER_ATLAS_KEY (32+ chars). Header only — never
        in a query string. Unset on the dispatcher, every route answers 503
        unavailable regardless of the key presented.
...
  responses:
    BadRequest:
      description: Payload failed validation — see details[].path
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: invalid_request
            message: Request failed validation
            details:
              - path: definition.sql
                message: SQL must be a SELECT or WITH statement
    Unauthorized:
      description: Missing, malformed or wrong key — same body for every reason
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: unauthorized
            message: Valid API key required
    NotFound:
      description: Unknown path, or a named resource that does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: not_found
            message: Variable "nope" does not exist
    Conflict:
      description: The write would duplicate a name that already exists
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: conflict
            message: Variable "available_credit" already exists
    RateLimited:
      description: More than 60 requests per minute for one key — back off
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: rate_limited
            message: Rate limit is 60 requests per minute
    Unavailable:
      description: >
        Key unset (integration not enabled), no state database attached, or the
        state database unreachable where degrading was impossible
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: unavailable
            message: "Atlas API is not configured on this dispatcher (DISPATCHER_ATLAS_KEY is unset)"
...
  schemas:
...
    ApiErrorCode:
      type: string
      description: Stable machine-readable code — branch on this, never on prose
      enum:
        - invalid_request   # 400 payload failed validation
        - unauthorized      # 401 missing/malformed/wrong key
        - not_found         # 404 unknown path or resource
        - conflict          # 409 duplicate name
        - rate_limited      # 429 over 60 req/min/key
        - unavailable       # 503 key unset / no usable state DB
        - internal          # 500 unexpected
    ErrorDetail:
      type: object
      required: [path, message]
      properties:
        path:
          type: string
          description: Dotted field path, e.g. definition.api.headers.Authorization
        message:
          type: string
    ErrorEnvelope:
      type: object
      required: [error, message]
      properties:
        error: { $ref: "#/components/schemas/ApiErrorCode" }
        message:
          type: string
        details:
          type: array
          description: Present on invalid_request only — one entry per failing field
          items: { $ref: "#/components/schemas/ErrorDetail" }
...
    CallMetadataKey:
      type: object
      required: [key]
      properties:
        key:
          type: string
          maxLength: 64
          pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$"
          description: >
            Unique within the schema; `field` is reserved. An api variable
            attaching the schema uses it as `{{key.k}}` (the key's name) and
            `{{key.v}}` (its value) anywhere in its request.
        placeholder:
          type: string
          maxLength: 500
          description: A sample value shown in the UI — never sent. Must match `regex` when both are set.
        regex:
          type: string
          maxLength: 500
          description: Validates the value once values are supplied per send. Must compile.
    CallMetadataInput:
      type: object
      required: [name, keys]
      properties:
        name: { type: string, maxLength: 191, pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$" }
        keys:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: "#/components/schemas/CallMetadataKey" }
...
    CallMetadata:
      allOf:
        - $ref: "#/components/schemas/CallMetadataInput"
        - type: object
          properties:
            id: { type: string, description: What api variables reference — stable across renames }
            used_by:
              type: array
              items: { type: string }
              description: Api variables attaching this schema
            required_keys:
              type: array
              items: { type: string }
              description: >-
                Keys used as `{{key.v}}` by a variable attaching this schema as
                required. A dispatch missing one (or sending a value that fails
                the key's regex) skips that variable's API call and renders its
                fallback.
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            updated_by: { type: string, nullable: true }
...